Basic Skeleton Done
A Basic Skeleton Done for the RL Training Environment + A tutorial for reference has been added as well from the Gymnasium API Docs,

However several errors are still in place so will need to be done and worked. Only reason it is being pushed currently is because I am still learning and this is a good point to see where I had gone wrong in my thinking. This Current version would understandably break most RL libraries such as Q-Agents, PPO, and CQL